Update Instruction
Added instructions:
Multiple items to perform the contracts instead of just one item. 
Can we have the user attached a file. It could be a csv, xls, text or pdf that includes all the contract to be negotiated in the event that this is a renegotiation.
Added disclosure that this is just a prediction and there other variables that can affect the results.